Move BuildSettings definitions to BuildSettings
